Rename generated_codegen to generated-codegen
