+++ .eslintignore
@@ -1,3 +1,3 @@
 /packages/core/src/generated/**/*.ts
-/packages/core/src/generated_codegen/**/*.ts
+/packages/core/src/generated-codegen/**/*.ts
 /dist/*
